update: guest

+++ view/client/cart/cart.php
@@ -20,27 +20,30 @@
                                 </tr>
                             </thead>
                             <tbody>
-                                <?php foreach ($cart as $car) : ?>
-                                    <tr>
-                                        <td class="product-thumbnail">
-                                            <a href="#"><img src="./images/product/<?= $car['pro_image'] ?>" width="90" height="100"></a>
-                                        </td>
-                                        <td class="product-name"><a href="#"><?= $car['pro_name'] ?></a></td>
-                                        <td class="product-price-cart"><span class="amount"><?= number_format($car['pro_var_sale_price'] * 1000) ?>đ</span></td>
-                                        <td><?= $car['var_color_name'] ?></td>
-                                        <td><?= $car['var_size_name'] ?></td>
-                                        <td class="product-quantity">
-                                            <div class="cart-plus-minus">
-                                                <input class="cart-plus-minus-box" type="text" name="quantity[<?= $car['cart_id'] ?>]" value="<?= $car['quantity'] ?>">
-                                            </div>
-                                        </td>
-                                        <!-- <td class="product-subtotal">$110.00</td> -->
-                                        <td class="product-remove">
-                                            <a href="#"><i class="fa fa-pencil"></i></a>
-                                            <a href="?act=delete_cart&cart_id=<?= $car['cart_id'] ?>"><i class="fa fa-times"></i></a>
-                                        </td>
-                                    </tr>
-                                <?php endforeach; ?>
+                                <?php
+                                if (!empty($_SESSION['user'])) {
+                                    foreach ($cart as $car) : ?>
+                                        <tr>
+                                            <td class="product-thumbnail">
+                                                <a href="#"><img src="./images/product/<?= $car['pro_image'] ?>" width="90" height="100"></a>
+                                            </td>
+                                            <td class="product-name"><a href="#"><?= $car['pro_name'] ?></a></td>
+                                            <td class="product-price-cart"><span class="amount"><?= number_format($car['pro_var_sale_price'] * 1000) ?>đ</span></td>
+                                            <td><?= $car['var_color_name'] ?></td>
+                                            <td><?= $car['var_size_name'] ?></td>
+                                            <td class="product-quantity">
+                                                <div class="cart-plus-minus">
+                                                    <input class="cart-plus-minus-box" type="text" name="quantity[<?= $car['cart_id'] ?>]" value="<?= $car['quantity'] ?>">
+                                                </div>
+                                            </td>
+                                            <!-- <td class="product-subtotal">$110.00</td> -->
+                                            <td class="product-remove">
+                                                <a href="#"><i class="fa fa-pencil"></i></a>
+                                                <a href="?act=delete_cart&cart_id=<?= $car['cart_id'] ?>"><i class="fa fa-times"></i></a>
+                                            </td>
+                                        </tr>
+                                <?php endforeach;
+                                } ?>
                             </tbody>
                         </table>
                     </div>
@@ -72,15 +75,17 @@
                             </div>
                         </div>
                         <div class="col-lg-4 col-md-12 mb-3">
-                            <div class="grand-totall">
-                                <div class="title-wrap">
-                                    <h4 class="cart-bottom-title section-bg-gary-cart">Hóa đơn</h4>
+                            <?php if (!empty($_SESSION['user'])) { ?>
+                                <div class="grand-totall">
+                                    <div class="title-wrap">
+                                        <h4 class="cart-bottom-title section-bg-gary-cart">Hóa đơn</h4>
+                                    </div>
+                                    <h5>Giá sản phẩm <span><?= number_format($sum * 1000) ?>đ</span></h5>
+                                    <h5>mã giảm giá <span>-<?= number_format(($_SESSION['totalCoupon'] ?? 0) * 1000) ?>đ</span></h5>
+                                    <h4 class="grand-totall-title">Tổng cộng <span><?= number_format(($sum - ($_SESSION['totalCoupon'] ?? 0)) * 1000) ?>đ</span></h4>
+                                    <a href="indext.php?act=checkout">Mua hàng</a>
                                 </div>
-                                <h5>Giá sản phẩm <span><?= number_format($sum * 1000) ?>đ</span></h5>
-                                <h5>mã giảm giá <span>-<?= number_format(($_SESSION['totalCoupon'] ?? 0) * 1000) ?>đ</span></h5>
-                                <h4 class="grand-totall-title">Tổng cộng <span><?= number_format(($sum - ($_SESSION['totalCoupon'] ?? 0)) * 1000) ?>đ</span></h4>
-                                <a href="indext.php?act=checkout">Mua hàng</a>
-                            </div>
+                            <?php } ?>
                         </div>
                 </form>
             </div>
